requirements.txt
flask==3.1.2
requests==2.32.5
lxml==6.0.1
python-dotenv==1.1.1
pandas>=1.5.0
numpy>=1.24.0